+++ Geometry_Engine/Compute/DistributeOutlines.cs
@@ -34,6 +34,90 @@ namespace BH.Engine.Geometry
         /**** Public Methods                            ****/
         /***************************************************/
 
+        /******************************************/
+        /****            IElement2D            ****/
+        /******************************************/
+
+        public static List<List<List<IElement1D>>> DistributeOutlines(this List<List<IElement1D>> outlines, bool canCutOpenings = true, double tolerance = Tolerance.Distance)
+        {
+            List<Tuple<PolyCurve, List<IElement1D>>> outlineCurves = new List<Tuple<PolyCurve, List<IElement1D>>>();
+            foreach (List<IElement1D> outline in outlines)
+            {
+                PolyCurve outlineCurve = new PolyCurve { Curves = outline.Select(x => x.IGeometry()).ToList() };
+                outlineCurves.Add(new Tuple<PolyCurve, List<IElement1D>>(outlineCurve, outline));
+            }
+
+            outlineCurves.Sort(delegate (Tuple<PolyCurve, List<IElement1D>> t1, Tuple<PolyCurve, List<IElement1D>> t2)
+            {
+                return t1.Item1.Area().CompareTo(t2.Item1.Area());
+            });
+            outlineCurves.Reverse();
+
+            List<Tuple<PolyCurve, List<IElement1D>, bool>> outlinesByType = new List<Tuple<PolyCurve, List<IElement1D>, bool>>();
+            foreach (Tuple<PolyCurve, List<IElement1D>> o in outlineCurves)
+            {
+                bool assigned = false;
+                for (int i = outlinesByType.Count - 1; i >= 0; i--)
+                {
+                    if (outlinesByType[i].Item1.IsContaining(o.Item1, true, tolerance))
+                    {
+                        if (canCutOpenings || i == 0)
+                            outlinesByType.Add(new Tuple<PolyCurve, List<IElement1D>, bool>(o.Item1, o.Item2, !outlinesByType[i].Item3));
+
+                        assigned = true;
+                        break;
+                    }
+                }
+
+                if (!assigned)
+                    outlinesByType.Add(new Tuple<PolyCurve, List<IElement1D>, bool>(o.Item1, o.Item2, true));
+            }
+
+            List<Tuple<PolyCurve, List<IElement1D>>> panelOutlines = outlinesByType.Where(x => x.Item3 == true).Select(x => new Tuple<PolyCurve, List<IElement1D>>(x.Item1, x.Item2)).ToList();
+            List<Tuple<PolyCurve, List<IElement1D>>> panelOpenings = outlinesByType.Where(x => x.Item3 == false).Select(x => new Tuple<PolyCurve, List<IElement1D>>(x.Item1, x.Item2)).ToList();
+            return panelOutlines.DistributeOpenings(panelOpenings, tolerance);
+        }
+
+        /***************************************************/
+
+        [DeprecatedAttribute("2.3", "Replaced with the same method taking an extra canCutOpenings argument.", null, "DistributeOutlines")]
+        public static List<List<List<IElement1D>>> DistributeOutlines(this List<List<IElement1D>> outlines, double tolerance = Tolerance.Distance)
+        {
+            return outlines.DistributeOutlines(true, tolerance);
+        }
+
+
+        /***************************************************/
+        /**** Private Methods                           ****/
+        /***************************************************/
+
+        private static List<List<List<IElement1D>>> DistributeOpenings(this List<Tuple<PolyCurve, List<IElement1D>>> panels, List<Tuple<PolyCurve, List<IElement1D>>> openings, double tolerance)
+        {
+            double sqTolerance = tolerance * tolerance;
+            List<List<Tuple<PolyCurve, List<IElement1D>>>> result = new List<List<Tuple<PolyCurve, List<IElement1D>>>>();
+            foreach (Tuple<PolyCurve, List<IElement1D>> panel in panels)
+            {
+                result.Add(new List<Tuple<PolyCurve, List<IElement1D>>> { panel });
+            }
+            result.Reverse();
+
+            foreach (Tuple<PolyCurve, List<IElement1D>> opening in openings)
+            {
+                for (int i = 0; i < result.Count; i++)
+                {
+                    if (result[i][0].Item1.IsContaining(opening.Item1, true, tolerance) && result[i][0].Item1.Area() - opening.Item1.Area() > sqTolerance)
+                    {
+                        result[i].Add(opening);
+                        break;
+                    }
+                }
+            }
+
+            return result.Select(x => x.Select(y => y.Item2).ToList()).ToList();
+        }
+
+        /***************************************************/
+
         [DeprecatedAttribute("Moved to Common_Engine")]
         public static List<List<Polyline>> DistributeOutlines(this List<Polyline> outlines, double tolerance = Tolerance.Distance)
         {
