+++ Geometry_Engine/Create/NewElement0D.cs
@@ -19,11 +19,10 @@
  * You should have received a copy of the GNU Lesser General Public License     
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
-
-using BH.oM.Common;
+ 
 using BH.oM.Geometry;
 
-namespace BH.Engine.Common
+namespace BH.Engine.Geometry
 {
     public static partial class Create
     {
