+++ Geometry_Engine/Modify/SetGeometry.cs
@@ -94,6 +94,27 @@ namespace BH.Engine.Geometry
             return clone;
         }
 
+        /******************************************/
+        /****            IElement0D            ****/
+        /******************************************/
+
+        public static IElement0D ISetGeometry(this IElement0D element0D, Point point)
+        {
+            return Reflection.Compute.RunExtensionMethod(element0D, "SetGeometry", new object[] { point }) as IElement0D;
+        }
+
+
+        /******************************************/
+        /****            IElement1D            ****/
+        /******************************************/
+
+        public static IElement1D ISetGeometry(this IElement1D element1D, ICurve curve)
+        {
+            return Reflection.Compute.RunExtensionMethod(element1D, "SetGeometry", new object[] { curve }) as IElement1D;
+        }
+
+        /******************************************/
+
         /***************************************************/
         /****              Interface Methods            ****/
         /***************************************************/
